requirements.txt
fastapi[standard]
uvicorn[standard]
motor
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-genai
pdfplumber
weasyprint
playwright
jinja2
pydantic[email]
slowapi
bleach
structlog
supabase
httpx
pytest
pytest-asyncio

# RAG / Langchain dependencies
langchain>=0.3.1
langchain-google-genai==2.0.0
langchain-community>=0.3.1
